chore: moves Confirmations specs to tests
