+++ tests/smoke/confirmations/send/send-solana-token.spec.ts
@@ -1,11 +1,11 @@
-import SendView from '../../pages/Send/RedesignedSendView';
-import SolanaTestDApp from '../../pages/Browser/SolanaTestDApp';
-import TokenOverview from '../../pages/wallet/TokenOverview';
-import WalletView from '../../pages/wallet/WalletView';
-import { SmokeConfirmations } from '../../tags';
-import { loginToApp } from '../../viewHelper';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
+import SendView from '../../../../e2e/pages/Send/RedesignedSendView';
+import SolanaTestDApp from '../../../../e2e/pages/Browser/SolanaTestDApp';
+import TokenOverview from '../../../../e2e/pages/wallet/TokenOverview';
+import WalletView from '../../../../e2e/pages/wallet/WalletView';
+import { SmokeConfirmations } from '../../../../e2e/tags';
+import { loginToApp } from '../../../../e2e/viewHelper';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
+import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
 
 const RECIPIENT = '4Nd1mZyJY5ZqzR3n8bQF7h5L2Q9gY1yTtM6nQhc7P1Dp';
