+++ tests/smoke/confirmations/send/send-tron-token.spec.ts
@@ -1,10 +1,10 @@
-import SendView from '../../pages/Send/RedesignedSendView';
-import TokenOverview from '../../pages/wallet/TokenOverview';
-import WalletView from '../../pages/wallet/WalletView';
-import { SmokeConfirmations } from '../../tags';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { loginToApp } from '../../viewHelper';
+import SendView from '../../../../e2e/pages/Send/RedesignedSendView';
+import TokenOverview from '../../../../e2e/pages/wallet/TokenOverview';
+import WalletView from '../../../../e2e/pages/wallet/WalletView';
+import { SmokeConfirmations } from '../../../../e2e/tags';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
+import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
+import { loginToApp } from '../../../../e2e/viewHelper';
 
 describe(SmokeConfirmations('Send TRX token'), () => {
   it.skip('shows insufficient funds', async () => {
