+++ tests/smoke/confirmations/transactions/7702/withDelegatorContracts.json
