+++ tests/smoke/confirmations/transactions/gas-fee-tokens-eip-7702-sponsored.spec.ts
@@ -1,32 +1,32 @@
-import FixtureBuilder from '../../../../tests/framework/fixtures/FixtureBuilder';
-import FooterActions from '../../../pages/Browser/Confirmations/FooterActions';
-import SendView from '../../../pages/Send/RedesignedSendView';
-import TabBarComponent from '../../../pages/wallet/TabBarComponent';
-import WalletView from '../../../pages/wallet/WalletView';
+import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
+import FooterActions from '../../../../e2e/pages/Browser/Confirmations/FooterActions';
+import SendView from '../../../../e2e/pages/Send/RedesignedSendView';
+import TabBarComponent from '../../../../e2e/pages/wallet/TabBarComponent';
+import WalletView from '../../../../e2e/pages/wallet/WalletView';
 import {
   Assertions,
   LocalNode,
   LocalNodeType,
   Utilities,
-} from '../../../../tests/framework';
-import { SmokeConfirmations } from '../../../tags';
-import { AnvilPort } from '../../../../tests/framework/fixtures/FixtureUtils';
-import { loginToApp } from '../../../viewHelper';
-import { withFixtures } from '../../../../tests/framework/fixtures/FixtureHelper';
-import RowComponents from '../../../pages/Browser/Confirmations/RowComponents';
-import { AnvilManager, Hardfork } from '../../../../tests/seeder/anvil-manager';
+} from '../../../framework';
+import { SmokeConfirmations } from '../../../../e2e/tags';
+import { AnvilPort } from '../../../framework/fixtures/FixtureUtils';
+import { loginToApp } from '../../../../e2e/viewHelper';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
+import RowComponents from '../../../../e2e/pages/Browser/Confirmations/RowComponents';
+import { AnvilManager, Hardfork } from '../../../seeder/anvil-manager';
 import {
   setupMockRequest,
   setupMockPostRequest,
-} from '../../../../tests/api-mocking/helpers/mockHelpers';
-import { SIMULATION_ENABLED_NETWORKS_MOCK } from '../../../../tests/api-mocking/mock-responses/simulations';
-import { setupRemoteFeatureFlagsMock } from '../../../../tests/api-mocking/helpers/remoteFeatureFlagsHelper';
-import { remoteFeatureEip7702 } from '../../../../tests/api-mocking/mock-responses/feature-flags-mocks';
+} from '../../../api-mocking/helpers/mockHelpers';
+import { SIMULATION_ENABLED_NETWORKS_MOCK } from '../../../api-mocking/mock-responses/simulations';
+import { setupRemoteFeatureFlagsMock } from '../../../api-mocking/helpers/remoteFeatureFlagsHelper';
+import { remoteFeatureEip7702 } from '../../../api-mocking/mock-responses/feature-flags-mocks';
 import { Mockttp } from 'mockttp';
 import {
   TRANSACTION_RELAY_STATUS_NETWORKS_MOCK,
   TRANSACTION_RELAY_SUBMIT_NETWORKS_MOCK,
-} from '../../../../tests/api-mocking/mock-responses/transaction-relay-mocks';
+} from '../../../api-mocking/mock-responses/transaction-relay-mocks';
 import { RelayStatus } from '../../../../app/util/transactions/transaction-relay';
 
 const TRANSACTION_UUID_MOCK = '1234-5678';
