+++ tests/smoke/confirmations/transactions/transaction-pay.spec.ts
@@ -1,28 +1,28 @@
-import { withFixtures } from '../../../../tests/framework/fixtures/FixtureHelper';
-import FixtureBuilder from '../../../../tests/framework/fixtures/FixtureBuilder';
-import { SmokeConfirmations } from '../../../tags';
-import { loginToApp } from '../../../viewHelper';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
+import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
+import { SmokeConfirmations } from '../../../../e2e/tags';
+import { loginToApp } from '../../../../e2e/viewHelper';
 import {
   remoteFeatureEip7702,
   remoteFeatureFlagPredictEnabled,
-} from '../../../../tests/api-mocking/mock-responses/feature-flags-mocks';
+} from '../../../api-mocking/mock-responses/feature-flags-mocks';
 import { Mockttp } from 'mockttp';
-import { setupRemoteFeatureFlagsMock } from '../../../../tests/api-mocking/helpers/remoteFeatureFlagsHelper';
-import { POLYMARKET_COMPLETE_MOCKS } from '../../../../tests/api-mocking/mock-responses/polymarket/polymarket-mocks';
-import PredictAddFunds from '../../../../tests/page-objects/Predict/PredictAddFunds';
+import { setupRemoteFeatureFlagsMock } from '../../../api-mocking/helpers/remoteFeatureFlagsHelper';
+import { POLYMARKET_COMPLETE_MOCKS } from '../../../api-mocking/mock-responses/polymarket/polymarket-mocks';
+import PredictAddFunds from '../../../page-objects/Predict/PredictAddFunds';
 import {
   mockRelayQuote,
   mockRelayStatus,
-} from '../../../../tests/api-mocking/mock-responses/transaction-pay';
+} from '../../../api-mocking/mock-responses/transaction-pay';
 import { CHAIN_IDS } from '@metamask/transaction-controller';
-import TransactionPayConfirmation from '../../../pages/Confirmation/TransactionPayConfirmation';
-import FooterActions from '../../../pages/Browser/Confirmations/FooterActions';
-import { Gestures } from '../../../../tests/framework';
-import TransactionDetailsModal from '../../../pages/Transactions/TransactionDetailsModal';
-import TabBarComponent from '../../../pages/wallet/TabBarComponent';
-import WalletActionsBottomSheet from '../../../pages/wallet/WalletActionsBottomSheet';
-import ActivitiesView from '../../../pages/Transactions/ActivitiesView';
-import PredictMarketList from '../../../../tests/page-objects/Predict/PredictMarketList';
+import TransactionPayConfirmation from '../../../../e2e/pages/Confirmation/TransactionPayConfirmation';
+import FooterActions from '../../../../e2e/pages/Browser/Confirmations/FooterActions';
+import { Gestures } from '../../../framework';
+import TransactionDetailsModal from '../../../../e2e/pages/Transactions/TransactionDetailsModal';
+import TabBarComponent from '../../../../e2e/pages/wallet/TabBarComponent';
+import WalletActionsBottomSheet from '../../../../e2e/pages/wallet/WalletActionsBottomSheet';
+import ActivitiesView from '../../../../e2e/pages/Transactions/ActivitiesView';
+import PredictMarketList from '../../../page-objects/Predict/PredictMarketList';
 
 describe(SmokeConfirmations('Transaction Pay'), () => {
   it('deposits to predict balance', async () => {
